GNCTestServer.jl

Test server for the Guidance, Navigation, and Control system of the PyCubed-Mini Sattelite

Package Features

Satelite dynamics, with magnetorquers

Server not included

Intended for use with the Pycubed-Mini emulation driver

Support for this server has not yet been created

Function Documentation

GNCTestServer.L — Method

L(q)

  Converts a scalar-first unit quaternion into the left-side matrix for 
quaternion multiplication, as described in "Planning with Attiude" (Jackson)

Arguments:
 - q:  A scalar-first unit quaternion                         |  [4,]

Returns: 
 - M:  Left-side matrix representing the given quaternion     |  [4, 4]

GNCTestServer.R — Method

R(q)

  Converts a scalar-first unit quaternion into the right-side matrix for 
quaternion multiplication, as described in "Planning with Attiude" (Jackson)

Arguments:
 - q:  A scalar-first unit quaternion                         |  [4,]

Returns: 
 - M:  Right-side matrix representing the given quaternion     |  [4, 4]

GNCTestServer.accel_perturbations — Method

Generates the acceleration for a spacecraft in LEO. Accounts for a variety of factors, including spherical harmonics, atmospheric drag, SRP, and thirdbody from sun and moon

ForwardDiff friendly (written by Kevin)

GNCTestServer.control_step — Method

Updates the free parameters given the state and time, then calls the control function to compute the control input.

Arguments:

state: Current state of the system, as a State struct | State

params: Free parameters of the system, as a Parameters struct | Parameters

control: Control function of state, parameters, and time | Function

t: Current time, as an Epoch struct | Epoch

Returns:

control: Control input of the Satellite | Control

GNCTestServer.default_log_init — Method

Default function to initialize the log, only logs angular velocity and magnitude.

Arguments:

state: Current state of the system, as a State struct | State

iterations: Number of iterations the simulation is run | Scalar

Returns:

hist: Initialized log of the simulation | Matrix

GNCTestServer.default_log_step — Method

Default function to update the log, only logs angular velocity and magnitude.

Arguments:

hist: Log of the simulation | Matrix

state: Current state of the system, as a State struct | State

i: Current iteration of the simulation | Scalar

GNCTestServer.dynamics — Method

Returns the derivative of the state, given the current state, parameters, control input, and time.

Arguments:

state: the current state of the spacecraft | State

parameters: the parameters of the spacecraft | Parameters

control: the control input to apply to the spacecraft | Control

t: the current time | Epoch

Returns:

state: the derivative of the state | State

GNCTestServer.hat — Method

hat(v)

Converts a 3-element vector into a cross-product matrix.

Arguments:
 - v:  3-element vector                   |  [3,] 

Returns:
 - M:  A [3 × 3] skew-symmetric matrix    |  [3, 3]

GNCTestServer.initialize_orbit — Method

Initializes a random, viable orbit given a few different terms, usually a position 'r' in Cartesian coordinates. Initial velocity may be specified, but if specified it will not necessarily result in a stable orbit.

The initial starting position, velocity, semi-major axis, orientation, and angular velocity may be either specified or determined randomly.

Arguments:

r: (Optional) Height above ground that the satellite will start its orbit at | Scalar

v: (Optional) Magnitude of initial velocity | Scalar

a: (Optional) Semi-major axis | Scalar

q: (Optional) Initial attitude, as a unit quaternion | [4,]

ω: (Optional) Initial angular velocity | [3,]

Returns:

x: Initial state, as (r, v, q, ω) | State

GNCTestServer.integrate_state — Method

Integrate the state forward by dt seconds, using the given control input.

Arguments

state: the current state of the spacecraft | State

parameters: the parameters of the spacecraft | Parameters

control: the control input to apply to the spacecraft | Control

t: the current time | Epoch

dt: the time step to integrate forward by | Float64

Returns

state: the state of the spacecraft after the integration | State

GNCTestServer.qdot — Method

qdot(q, ω)

Provides the derivative of a quaternion, given the angular velocity 

Arguments:
 - q:  Scalar-first unit quaternion                              |  [4,]
 - ω:  Angular velocity                                          |  [3,]

Returns:
 - q̇:  Derivative of attitude, parameterized as a quaternion     |  [4,]

GNCTestServer.quaternionToMatrix — Method

quaternionToMatrix (q) Arguments: - q: Scalar-first unit quaternion | [4,]

Returns:
 - Q: Rotation matrix representing the same rotation

GNCTestServer.rk4 — Method

Rung-Kutta 4th order integrator

GNCTestServer.sim — Function

Runs the simulation, given a control function. Optionally takes in functions to initialize and update the log.

Arguments:

control_fn: Function to compute control input, given state, params, and time | Function

log_init: (Optional) Function to initialize the log, given the number of steps | Function

log_step: (Optional) Function to update the log, given the current state | Function

Returns:

hist: Generated log of the simulation

GNCTestServer.sim_step — Method

steps the simulation forward by one time step, dt. given the current state, free parameters, control function, and time.

Arguments:

state: Current state of the system, as a State struct | State

params: Free parameters of the system, as a Parameters struct | Parameters

control: Control input of the Satellite | Control

t: Current time, as an Epoch struct | Epoch

dt: Time step, in seconds | Scalar

Returns:

state: Updated state of the system, as a State struct | State

t: Updated time, as an Epoch struct | Epoch